# LMSproj/DTO/AddBookDto.cs
namespace LMSproj.DTO
{
    


    public class AddBookDto

    {

        public string Title { get; set; }

        public string Author { get; set; }

        public string ISBN { get; set; }

        public int TotalCopies { get; set; }

        public string Category { get; set; }

    }



    
}